requirements.txt
streamlit==1.29.0
pandas==2.2.0
plotly==5.21.0
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.4.2
textblob==0.18.0
nltk==3.8.1
pytesseract==0.3.10
bcrypt==4.1.2
streamlit-option-menu==0.3.6
Pillow==10.2.0
